pyproject.toml
[project]
name = "elastic-product-search-lab"
version = "0.1.0"
description = "A compact Elasticsearch product-search lab for relevance, ingestion, and hybrid retrieval experiments."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "elasticsearch",
  "python-dotenv",
  "pytest",
  "pydantic"
]

[project.optional-dependencies]
esci = [
  "pandas",
  "pyarrow"
]
vector = [
  "sentence-transformers",
  "scikit-learn",
  "rank-bm25"
]
kafka = [
  "confluent-kafka"
]

[tool.pytest.ini_options]
testpaths = ["tests/python"]
pythonpath = ["."]
markers = [
  "integration: tests that require local services such as Elasticsearch"
]
